Added the Navigation Bar component
This component is made for the users to be able to navigate throught the most important links and services of the website, such as projects, courses etc. The dropdown menus are capable of closing the previous one to enhance  the user experience and to be able to see the other dropdown links without overlapping the other. This navigation bar is responsive to the smallest devices, such as Samsung Galaxy Fold, and older iPhone devices.